Rename of workflow_ors, separate commit for easier CR
