+++ packages/jira-adapter/src/adapter.ts
@@ -120,7 +120,7 @@ import accountInfoFilter from './filters/account_info'
 import deployPermissionSchemeFilter from './filters/permission_scheme/deploy_permission_scheme_filter'
 import scriptRunnerWorkflowFilter from './filters/script_runner/workflow_filter'
 import pluginVersionFliter from './filters/data_center/plugin_version'
-import scriptRunnerWorkflowOrFilter from './filters/script_runner/workflow_ors'
+import scriptRunnerWorkflowListsFilter from './filters/script_runner/workflow_lists_parsing'
 import storeUsersFilter from './filters/store_users'
 
 const {
@@ -202,7 +202,7 @@ export const DEFAULT_FILTERS = [
   fieldConfigurationSchemeFilter,
   scriptRunnerWorkflowFilter,
   // must run after scriptRunnerWorkflowFilter
-  scriptRunnerWorkflowOrFilter,
+  scriptRunnerWorkflowListsFilter,
   userFilter,
   forbiddenPermissionSchemeFilter,
   jqlReferencesFilter,
