+++ packages/jira-adapter/src/filters/script_runner/workflow_lists_parsing.ts
