+++ packages/jira-adapter/test/filters/script_runner/workflow_mail_lists.test.ts
@@ -17,7 +17,7 @@ import { filterUtils } from '@salto-io/adapter-components'
 import { InstanceElement, toChange } from '@salto-io/adapter-api'
 import _ from 'lodash'
 import { createEmptyType, getFilterParams, mockClient } from '../../utils'
-import orFilter, { MAIL_LISTS_FIELDS } from '../../../src/filters/script_runner/workflow_ors'
+import orFilter, { MAIL_LISTS_FIELDS } from '../../../src/filters/script_runner/workflow_lists_parsing'
 import { WORKFLOW_TYPE_NAME } from '../../../src/constants'
 import { getDefaultConfig } from '../../../src/config/config'
